examples/ArtLife_v1/requirements.txt
numpy==1.23.5
Pillow==9.3.0
pygame==2.6.1